\name{Mortality}
\alias{Mortality}
\docType{data}
\title{
Swedish male mortality data from the HMD.
}
\description{
Four different mortality items are taken from the Human Mortality Database to demonstrate the functionality of various (ad-hoc) functions for splitting Mx data into Lexis triangles. Each of the four elements is a matrix with years going across columns (1900-1989) and ages going down rows (0-100). \code{mxmat} are Mx values, \code{Dxmat} are death counts, \code{Expmat} are population exposures, \code{lexismat} are deaths in lexis triangles. These were all extracted using \code{HMDget()} and have been unaltered.
}
\usage{data(Mortality)}
\format{
  The format is:
List of 4
 $ mxmat   : num [1:100, 1:90] 0.11951 0.02986 0.01558 0.01156 0.00917 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:100] "0" "1" "2" "3" ...
  .. ..$ : chr [1:90] "1900" "1901" "1902" "1903" ...
 $ Dxmat   : num [1:100, 1:90] 7686 1804 928 673 528 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:100] "0" "1" "2" "3" ...
  .. ..$ : chr [1:90] "1900" "1901" "1902" "1903" ...
 $ Expmat  : num [1:100, 1:90] 64318 60427 59568 58183 57651 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:100] "0" "1" "2" "3" ...
  .. ..$ : chr [1:90] "1900" "1901" "1902" "1903" ...
 $ lexismat: num [1:200, 1:90] 5353 2334 1023 782 479 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:200] "0_lower" "0_upper" "1_lower" "1_upper" ...
  .. ..$ : chr [1:90] "1900" "1901" "1902" "1903" ...
}
\source{
Human Mortality Database. University of California, Berkeley (USA), and Max Planck Institute for Demographic Research (Germany). Available at www.mortality.org or www.humanmortality.de (data downloaded on 28-7-2011). 
}
\examples{
# see examples for the four different functions for triangle interpolation.
}
\keyword{datasets}
